migrate: to ESLint Flat Config - https://eslint.org/docs/v8.x/use/configure/migration-guide - https://typescript-eslint.io/blog/announcing-typescript-eslint-v7/#new-features---flat-config-support - https://sveltejs.github.io/eslint-plugin-svelte/user-guide/#new-config-eslint-config-js - https://zenn.dev/babel/articles/eslint-flat-config-for-babel#flat-config-への移行-tips
